refactor: Replace components
Replace BookList, BookListItem, CartTable from /page-components to
/components.
It necessary to easy imports and work with them.
