+++ src/components/book-list-item/index.js
