--- src/components/pages/page-components/index.js
@@ -1,9 +1,0 @@
-import BookList from './book-list';
-import BookListItem from './book-list-item';
-import CartTable from './cart-table';
-
-export {
-  BookList,
-  BookListItem,
-  CartTable,
-};
